ChainSync client: simplify a correctness argument
We introduce a small generic helper function `cross` whose correctness follows
immediately from the guarantees of `AF.intersect`; removing the need for ad-hoc
inline justifications.

Note that this change improves sharing when multiple ChainSync clients maintain
candidate fragments that are (small) extensions of our selection. A priori, this
is contrary to our design principle of "average case = worst case", but we think
the effect is sufficiently small compared to the overall resource usage of a
node in order to influence an SPO to downgrade their node due to this.

Concretely, a Cardano header has a maximum size of `1.1 kB`, and `ourPre`
contains at most `k = 2160` headers, so we get a maximum size decrease of

    1.1 kB × 2160 × 100 = 237.6 MB

very conservatively assuming 100 ChainSync clients.

Co-authored-by: Alexander Esgen <alexander.esgen@iohk.io>