fix bug with registerGenesisStaking
The function 'registerGenesisStaking' is a hacky function, to be used
only in testing, which sets up the ledger state in a way that is
fraught with dragons (since it bypasses our usual protections).

Indeed, a recent change to the ledger resulted in breaking this
function. In particular, the ledger now memoizes the per-stake-pool
stake distribution in a way that needed to be accounted for inside
'registerGenesisStaking'.

See https://github.com/input-output-hk/cardano-ledger/blob/cffa75fdbd800cda60997791e51bf02f2af0c42b/docs/adr/2022-12-12_007-optimize-ledger-view.md
for more information on the changes in thet ledger.