fix bug with registerGenesisStaking
The function 'registerGenesisStaking' is a hacky function, to be used
only in testing. It sets up the ledger state in a way that is
fraught with dragons (since it bypasses our usual protections).

Indeed, a recent change to the ledger resulted in breaking this
function. In particular, the ledger now memoizes the per-stake-pool
stake distribution in a way that needed to be accounted for inside
'registerGenesisStaking'.

See ADR-7 in the ledger repository
(https://github.com/input-output-hk/cardano-ledger) for more details.

This fix is required to bring back the TICKF optimizations (see ADR-7)
in the ledger packages and have all the tests in the cardano-node
repository pass.