Add a `ci` workflow
This `ci` replaces the following workflows:

- `pages`
- `publish-mempool-benchmarks`
- `haskell-ci`

We build and deploy the GitHub pages on the same workflow because we
need to publish the haddocks, which depend on the Cabal build.
Similarly, once we publish the benchmarks in our site we will need to
share the benchmarks results with the job that publishes the GitHub
pages. Sharing artifacts or caches between GitHub workflows is not
supported, hence we need to put all these jobs in the same workflow.

This commit also removes support for publishing the benchmark plots. The
action we used
https://github.com/benchmark-action/github-action-benchmark is not
compatible with the new method we use to publish GitHub pages.