Add `strict-mvar` dependency and `StrictMVar` with NoThunks invariants.
We depend on an unreleased version of `strict-mvar` and we create a new
module for `StrictMVar`s with `NoThunks` invariants. We also add a flag
`checkmvarnothunks` that acts as a switch for whether we want to run the
invariant checks or not. By default, these checks are off, since we do
not want to perform these checks in production.