Fix `pure @(NonEmpty xs)` implementation
The previous implementation violated e.g. the Identity law, e.g. for

    v :: NonEmpty [String, Bool, Char] ()
    v = NonEmptyCons () (NonEmptyOne ())

we have

    pure id <*> v = NonEmptyOne ()

which is not equal to `v`.

The correct implementation is similar to the `Applicative ZipList` instance: We
provide an `a` for every `xs`.

Luckily, `pure @(NonEmpty xs)` was not used anywhere in our code base so far.