Process HF queries at HF level
HF queries are usually forwarded to the underlying single era block,
however due to the ledger tables in the backing store being an instance
of the HardFork LedgerTables, this becomes very expensive.

This commit splits the queries by their footprint as it was done in the
past, and defines the `QueryLedgerHF` class to process them at the hard
fork level.