Byron: move tip block number into `ByronTransition`
Initial motivation for this is the need (in a later commit) to implement the
translation of the last *ticked* Byron ledger state to the first Shelley state,
which becomes possible as `ByronTransition` is part of the ticked Byron ledger
state.

Apart from the translation to Shelley, this tip block number is only used to
check whether an update proposal became stable, so it makes sense to make it
part of `ByronTransition` in any case.