(T)Praos `ChainDepState`: slot now tracks where we last ticked to
In a single-era context, this change makes no difference for the slot field in
the unticked `ChainDepState (T)Praos`, as the slot we last ticked to will always
be the slot of the last applied block.

However, in a later commit, we will change how the HFC does cross-era ticking,
namely by first ticking just across the epoch boundary using the old era's
logic, then translating, and then ticking to the target slot using the new era's
logic.

 - Without this change, as the slot in the unticked `ChainDepState` is not
   update by ticking, we would conclude twice that we are ticking across an
   epoch, which is incorrect.

 - With this change, only the first tick considers itself to cross an epoch
   boundary.

Hence, after the translation step, the slot in the unticked `ChainDepState` will
*not* contain the slot of the last applied block, but rather the slot just after
the era boundary (as we ticked to that slot in the first tick). Note that this
is purely an intermediary state in the HFC.