`tickChainDepState`: factor through new `HorizonView` type
Context: This commit enables changing the HFC protocol logic for ticking.

This change is based on the observation that all single-era `ConsensusProtocol`s
except TPraos do not actually need the `LedgerView` when ticking at all; they
just store it in the `Ticked ChainDepState` such that it can be used by other
functions from `ConsensusProtocol`, such as `updateChainDepState` and
`checkIsLeader`.

Hence, we can introduce a new type family `HorizonView` (that will be set to
`()` for all single-era protocols except TPraos), and decompose
`tickChainDepState` into `tickChainDepState_` (which now only takes a horizon
view instead of a ledger view) and `projectHorizonView` (which allows one to
always get a horizon view from a ledger view).

TPraos and the HardForkProtocol are just minimally modified to make them
compile; they will be adapted in subsequent commits.

Co-authored-by: Nicolas Frisby <nick.frisby@iohk.io>