consensus: add and use PreparedChainDepState and PreparedHeaderState
The key benefit is that the Ticked ChainDepState types no longer need to carry
the LedgerView.

The one downside is that this introduces an error case in the
HardForkCombinator, since PreparedChainDepState will hold a Ticked LedgerView
and a Ticked ChainDepState side-by-side there, and its only guaranteed that
those be in the same era (ie the same summand in the telescope) by the
invariant on PreparedChainDepState.

We consider the resulting exactness of the Ticked ChainDepState to be worth the
impossible alternative in distribPreparedChainDepState. In particular, it will
unblock another refinement we're working on related to ticking ChainDepStates
across era transitions.

P.S. - This is effectively reverting 16d61bf708129ada44ec0abaf15f219087509e36.