Separate block tree analysis and abstract point schedule dispatch
`PeerSimulator.ScheduledChainSyncServer` now only implements the
boilerplate of fetching the next item from an STM action (the queue) and
maintaining it as shared state between the handlers, in a
`ChainSyncServer`.

`PeerSimulator.Handlers` contains the logic that determines the
intersection and next header from the block tree.

`PeerSimulator.Run` contains only the scheduler, operating on the minimal
set of resources.

`PeerSimulator.Resources` contains the data types and resource
acquisition constructors.