Save the prefix from the trunk instead of the branch
The precondition ensures that the prefix of the trunk contains the prefix of the branch, so no information would be lost there. When manipulating the prefix of a branch later, one of course expects it to be the prefix from the root of the tree (aka anchor of the trunk) and not from a random point depending on how the branch was added.