Make BlockSend a scheduled handler
The previous implementation of the scheduled BlockFetch server used BP
as an indicator of whether to start sending a range of blocks.

That meant that when the client requested some blocks that were within
bounds of HP, but BP wasn't beyond the _latest_ block of the requested
range, we would stall until a later tick, and send the entire range of
blocks in that tick.

We now decided that it would be a better manifestation of the test plan
if BP controlled whether an _individual_ block is sent within a given
tick.
Therefore, we now use HP to transition to sending a batch, then hook
into the scheduler before sending each block.