Implement lightweight checkpointing
Lightweight checkpointing is a mechanism to ensure that new nodes
end up in the correct chain even when the chain is too sparse for
normal operation of Praos and Genesis.

The idea is to supply the new node with a list of block hashes that
should be present on the chain at specific block numbers. This list
is provided in the TopLevelConfig record, and it is used during
validation of headers in validateIfCheckpoint called by
validateEnvelope.

If the hashes of checkpoints don't match the hashes of a supplied
header for a given block number, then validation of the header fails.

The substance of the change is modifying TopLevelConfig and
validateEnvelope. Most other changes derive from these modifications.
There are three new unit tests of validateIfCheckpoint added in
ouroboros-consensus:test:consensus-test.