Ignore/fix broken links
Problem: several links in Markdown files are broken.

Solution:
1. Links in References.md, index.md and TechnicalReports.md
are invalid in the sense of Markdown and this repo, but
they work on the website, so they are explicitly ignored.
2. Glossary.md contains many links to anchors in the same file.
These anchors work differently in Markdown and Docusarus as can
be seen here:
https://ouroboros-consensus.cardano.intersectmbo.org/docs/for-developers/Glossary/
Ideally, they should be fixed to work on the resulting website,
in which case all of them would be broken in the sense of Markdown.
Fixing them for Docusaurus is out of scope of this commit,
but we force xrefcheck to ignore all links in that file in
preparation for that fix.
3. There is also an anchor link in VersioningSchemeDecision.md,
but this file is not used in the resulting website, so we fix
it according to Markdown rules.