In `SinglePeer`, differentiate between relative and absolute times
`SinglePeer` and `SinglePeer.Indices` use `DiffTime` for everything,
which is technically correct but which I find extremely confusing.
`si-timers` already provides a `Time` wrapper around `DiffTime`,
described as “a point in time in a monotonic clock.” I think it helps
reading the code and it catches bugs where we eg. attempt to add two
times that are both morally absolute.