Add `BlockSupportsDiffusionPipelining`
This commit allows to specify a block-specific criterion for when a header can
be pipelined, instead of always using a `SelectView`-based rule.

For now, all blocks (except the HFC) implement this new type class via
`SelectViewTentativeState`, retaining the previous behavior. A later commit will
change the behavior for all Shelley-based blocks starting in Conway.