Trigger ChainSel (and LoE update) whenever a candidate fragment changes
When a node gets stuck at the LoE limit for a while, it's possible that
all blocks up to the forecast horizon have been added to the ChainDB
before the LoE fragment can advance as a result of a killed peer.

Even if we trigger the LoE when a peer is killed, it's possible that the
chain cannot advance when the honest peer is behind on the headers and
the adversary already sent all the blocks and gets stuck at the forecast
horizon, therefore not being subject to timeouts.

So instead we trigger the LoE on every header.
In a conservative approach, we fetch all direct successors of each of
the chain's blocks and run ChainSel for them whenever a candidate
changes.
This has good average==worst case characteristics but is quite
expensive, so a follow-up change will improve its performance.

Co-authored-by: Torsten Schmits <git@tryp.io>