Add `ChainOrderConfig`
A follow-up commit will introduce the restricted VRF tiebreaker based on slot
distance for Praos. As pre-Conway diffusion pipelining relies on the chain order
being objective, we have to still use unrestricted VRF tiebreakers before
Conway.

Another option would be to allow to parameterize the Praos protocol by the
SelectView that should be used. However, that still wouldn't allow for a clean
way to pass in the maximum slot distance in which VRF tiebreakers should be
used.