Don't disconnect peers before they have a fair chance to send headers
Imagine two peers serving the following chain

k: 1
sgen: 2
  0 1 2
G---1-2

Say peer1 sent no headers yet and peer2 sent 2 headers.
The intersection of both is G, the density of peer2's chain is 2,
while the upperbound of the density of peer1 is also 2.

Before this commit, GDD would disconnect peer1 as it cannot improve
the density of peer2's chain. For this decision to be correct, however,
it is essential that both chains disagree after the intersection.

To know if the chains will dissagree we defer disconnecting peer1
until it declares to have no more headers, or until it sends one header
after the intersection which disagrees with peer2's chain.