Simplify the honest peer shrinking function
Modify the honest shrinking function by no longer speeding up the other
schedules when an honest tick is deleted. This simplifies a lot of code
in the `Shrinking` module, at the cost of no longer ensuring that shrunk
schedules preserve the overall order of events.

Additionally, we re-enable shrinking in CSJ tests

Also:
* Extend adversarial schedules when shrinking an honest one
* Document the cases when we don't use shrinking