Refactor `PointSchedule` to support test end time
This commit refactors the `PointSchedule` type from a `newtype` of
`Peers (PeerSchedule blk)` to a datatype containing this schedule as
well as an additional field `psMinEndTime`. This field describes a
minimal absolute time that the test must reach. If all ticks are
executed before this time is reached, an extra delay is inserted.

This allows getting rid of all the places in which we added extra ticks
only for the purpose of making the test run longer. At the cost of
complexifying a bit the implementation of point schedules, this makes
the semantics much clearer.

The fact that `PointSchedule` is now a datatype makes it easy to later
add other fields, for instance a field stating the initial tip points,
instead of having to add zero-duration ticks at the beginning to set
those up, or a field describing the origin of the absolute clock,
instead of shifting the tick times.

Co-authored-by: Nicolas “Niols” Jeannerod <nicolas.jeannerod@tweag.io>