db-truncater: make slot truncation constant-time
This also slightly changes the semantics: Previously, it would truncate to the
block with the largest slot that is smaller or equal to the slot argument. Now,
it will only truncate to the (non-EBB) block in exactly that slot, and fail
otherwise.

In fact, the new behavior more closely corresponds to the CLI description:

  --truncate-after-slot SLOT_NUMBER
                           The slot number of the intended new tip of the chain
                           after truncation