ImmutableDB: allow to lookup blocks by slot, leverage in db-{analyser,truncater} (#1143)
This PR adds a new function to the internal ImmutableDB API:
```haskell
-- | Get the hash of the block in the given slot. If the slot contains both
-- an EBB and a non-EBB, return the hash of the non-EBB.
getHashForSlot :: SlotNo -> m (Maybe (HeaderHash blk))
```
It is then used in two ways:

- In db-analyser: Many analyses do not actually need a ledger state to
perform their work. This PR adds support for avoiding this redundant and
long work on startup. However, the existing ImmutableDB streaming API
needs a point (previously read from the ledger state), not just a slot
to start streaming. This is where `getHashForSlot` comes in.

- In db-truncater: Previously, truncating after a slot took linear time
(by iterating over the entire database). With `getHashForSlot`, it is
easy to now change it to take constant time. Note that the behavior
changes slightly, see the commit message for details.