consensus: consolidate transaction limits in the mempool
Transaction size, block capacity, and mempool capacity are multi-dimensional
vectors (incl ExUnits, etc), instead of merely bytes: see `TxMeasure`.
`TxLimits` has fittingly been promoted from the `Capacity` module to the proper
`SupportsMempool` module, cannibalizing some methds from the
`LedgerSupporstMempool` class.

A transaction cannot be added if it would push any component of the size over
that component's capacity.

The capacity override is still only specified in byte size, but the value is
interpreted as a block count (rounded up).

Enforce block capacity _before_ the forging logic.
Now the forging logic simply includes whatever transactions its given, which is
reasonable and simpler. It's the NodeKernel logic that uses the mempool's
finger tree in order to slice an appropriately-sized prefix, which is then
passed to the now-dumb forging function.

Explicit attention is given to overflow and the DoS vector of providing an
erroneously massive tx that would block the mempool's growth, were it not for
the new guards.

Also, anachronistically use ConwayMeasure in Babbage, since it should have been
counting ref script bytes, as Conway now does.

Many comments are improved and also updated for the new scheme.