consensus: use Natural for ByteSize
At the cost of a single pointer indirection --- which we expect to be dwarfed
by crypto checks etc --- the Mempool is trivially innoculated against
overflows.

The worst-case is actually at most two pointers, and that can only happen when
a Word64 actually would have overflowed. This should never happen in practice
due to the Diffusion Layer's finitely-bounded message buffers (for the
TxSubmission mini protocol, in particular) (perhaps unless the node is severely
configured). But a minor performance hit seems a preferable worst-case compared
to crashes/confused conditionals/etc.