Implement the Cached index using TVars
In the past some nodes have been crashing because of a misalingment in the data in the Cached Index.
Ensuring that the changes are atomic via `stm` seems to solve the problem.

In terms of logic, this PR only moves the modifications into `atomically` blocks, the overall logic
is the same.