ChainSync client: allow to prevent historical rollbacks
We only enable this check when ChainSync Jumping is enabled, in order to prevent
adversarial peers that got disengaged after doing a historical rollback before
an acknowledged point from wasing CPU and network bandwidth.

On the other hand, we do not have to prevent historical rollbacks in Praos
mode (as all peers are trusted) or when CSJ is disabled (as the Limit on
Patience then prevents historical rollbacks from causing much additional load).