Set `retention-days` to `10` for benchmark artificats in GHA
A lifetime of only a single day is prohibitive. If the benchmark jobs are run 24 hours after the build+test jobs, then the benchmark jobs will fail because the artifact will have been deleted by then. A stop-gap solution is to rerun all jobs, but that's not very nice UX. In this commit, we change retention days to 10.