Don't use `ltcollapse` like it is a fold
I had initially decided it was best to replace uses of `ltcollapse` by some new
`ltfoldMap` function, but after some thinking it's best to instead use
`ltcollapse` as is, but removing the use of monoids there, which currently has
no effect.

The reason I think this is the right call is because ledger tables are currently
a single-constructor newtype, and they will be for at least a while. We do not
know what ledger tables will look like when we store more parts of the ledger
state, so let's cross that bridge when we get there.