CSJ: prefer the Objector when replacing a Dynamo
This patch was introduced in response to a CSJ test failure; the property
expects that no honest header will be downloaded multiple times. Shape of the
failing test:

- The first Dynamo was adversary 3.
- honest 3 became the Objector, and killed that Dynamo.
- The next Dynamo was adversary 1.
- honest 1 became the Objector, and killed that Dynamo.

As as a result, some honest headers were downloaded by both honest 1 and honest
3.

This commit also renames `electNewDynamo` to `backfillDynamo` in order to
empahsize that this is only called when the Dynamo is lost. In particular, it's
not called by `rotateDynamo`.