Add version 0.5 and its publishing date to the utxo-db-api doc
Soon, we'll publish a technical report that is related to lsm-tree, which refers
to version 0.5 of the "Storing the Cardano ledger state on disk: API design
concepts" document. Currently the latter only mentions version 0.6 on its title
page, but to reduce confusion with publishing dates, we'll also add version 0.5
to the list of published versions, with the appropriate date.

I'v added changelogs to the two `utxo-db` and `utxo-db-api` docs, to facilitate
the versioning.