UTXO-HD
This commit puts together all the changes for UTxO-HD. It is annoying that it is
such a big commit but due to the numerous rebases, conflicts arose at each
single commit thus we reverted to put together everything in one commit to
minimize the conflict-resolving work.

For a super high-level description of UTxO-HD, the changelogs give brief
indications. For a more low-level description, check the new website documents.

This commit is the result of the joint effort by many people that are or were
part of the Consensus team and neighbour teams that made contributions be those
in design, testing or implementing some parts of UTxO-HD, who I would like to
acknowledge their participation by adding them as co-authors here, in no
particular order.

Co-authored-by: Joris Dral <joris@well-typed.com>
Co-authored-by: Alexander Esgen <alexander.esgen@iohk.io>
Co-authored-by: Nick Frisby <nick.frisby@iohk.io>
Co-authored-by: Damian Nadales <damian.nadales@iohk.io>
Co-authored-by: Fraser Murray <fraser.murray@iohk.io>
Co-authored-by: Georgy Lukyanov <georgy.lukyanov@iohk.io>
Co-authored-by: Renate Eilers <renate.eilers@iohk.io>
Co-authored-by: Yogesh Sajanikar <yogesh.sajanikar@iohk.io>
Co-authored-by: Bart Frenk <bart.frenk@iohk.io>
Co-authored-by: Pawel Szulc <pawel.szulc@iohk.io>
Co-authored-by: Arnaud Bailly <arnaud.bailly@iohk.io>
Co-authored-by: Douglas Wilson <douglas@well-typed.com>
Co-authored-by: Duncan Coutts <duncan.coutts@iohk.io>
Co-authored-by: Alexey Kuleshevich <alexey.kuleshevich@iohk.io>
Co-authored-by: Jared Corduan <jared.corduan@iohk.io>