ChainSel: remove a redundant intersection computation
`Diff.diff` computes an intersection between two chains; which is a bit ironic
as the comment on `switchFollowerToFork` talks about passing in the intersection
point as an optimization as it has already been computed.

We can avoid easily this intersection computation by noting that the `chainDiff`
records (via `getRollback`) how many blocks at the end of `curChain` we are
orphaning by the currently performed chain switch.

Also add a type signature for clarity.