ChainDB: perform LedgerDB tasks in separate background thread
Primarily, this is an optimization to reduce the maximum memory usage (more
relevant with the in-memory backend) when pruning happens on garbage collection
instead of while adding new blocks to the LedgerDB, see the added commit and the
benchmark in the pull request. Previously, LedgerDB garbage collection happened
as part of VolatileDB garbage collection, which was intentionally rate-limited.

Also, it resolves the current (somewhat weird) behavior that we do not copy any
blocks to the ImmutableDB when we are taking a snapshot (which can take >2
minutes), and consequently also not garbage-collecting the VolatileDB.

It also synergizes with the planned feature to add a random delay when taking
snapshots.