LedgerDB.V2: make sure to actually close handles (#1516)
Closes #1515

The first three commits add a regression test by enrichting the existing
LedgerDB state machine test, making sure that
```
  total number of all opened handles
- total number of closed handles
= states on the LedgerSeq in the end
```
The fourth commit then fixes the bug, causing the regression test to
succeed.

Note that this bug fix reveals another bug #1551 (causing the
`mock-test` failure for `LeaderSchedule.simple convergence`), which we
silence by temporarily making the closing of an in-memory handle a no-op
(as it is not *necessary* anyways, the GC will still collect the
in-memory ledger state).