LedgerDB.V2: prevent race condition when snapshotting (#1555)
Closes #1551

The commits are intended to be reviewed individually.

- The first commit fixes the race condition described in #1551 by using
the `ldbOpenHandlesLock` to duplicate the handle when snapshotting.
- The second commit reverts 7900088a90ec341f563c909be2ca3af326f2ca5d
added in #1516, enabled by the first commit. Without the first commit,
this would cause test failures, see the commit description.
- The third and fourth commits are purely refactorings to be able to
reuse the pattern of "duplicating while holding a read lock".
- The fifth commit uses the same pattern for `takeSnapshotNOW`, a
testing function, see the commit description.
 - The last commit adds documentation for `ldbOpenHandlesLock`.