Properly cleanup forkers at the appropriate times (#1564)
# Description

- The backing store of the V1 LedgerDB was only tracked in the resource
registry if we were starting from Genesis. Now the backing store will be
properly tracked in the resource registry even when we start from a
snapshot.

- Closing the LedgerDB will no longer release all the open forkers, but
instead invalidate them by emptying the ldbForkers map, so that the only
possible operation that could be performed is closing them in the
LedgerDB clients, such as ChainSel or the forging loop.

- Closing the forker is idempotent, and it was performed both when
calling `forkerClose` as well as when the resource registry of the
LedgerDB client was going out of scope. Now, `forkerClose` will release
the resource from the registry so this won't run twice.